requirements.txt
numpy
pandas
scikit-learn==1.3.2
matplotlib
pillow
joblib
tqdm
gradio
opencv-python